requirements.txt
python-dateutil>=2.6.0
selenium>=3.4.3
six>=1.10.0
python-axolotl
cryptography
typing
numpy
python-magic
requests
werkzeug
flask
flask-socketio
pydub
emoji
python-telegram-bot